[framework] feat: Introduced new framework files for quantization and model workload tooling. [Models] feat: Add a series of models in main repository, thank to:

+++ models/models/DeepSeekR1/import-deepseek-r1.py
@@ -71,7 +71,9 @@ if model_path is None:
     model_path = "deepseek-ai/DeepSeek-R1-Distill-Qwen-1.5B"
 
 # Initialize the model from the specified model path.
-model = AutoModelForCausalLM.from_pretrained(model_path).eval()
+model = AutoModelForCausalLM.from_pretrained(
+    model_path, dtype=torch.float32
+).eval()
 model.config.use_cache = False
 
 # Initialize Dynamo Compiler with specific configurations as an importer.
